% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dkt_3d}
\alias{dkt_3d}
\title{Desikan-Killiany-Tourville 3D Atlas (Legacy)}
\usage{
dkt_3d()
}
\value{
A legacy ggseg3d atlas tibble.
}
\description{
Desikan-Killiany-Tourville 3D Atlas (Legacy)
}
\examples{
dkt_3d()
}
\references{
Klein A, Tourville J (2012).
  101 labeled brain images and a consistent human cortical labeling protocol.
  *Frontiers in Neuroscience*, 6:171.
  \doi{10.3389/fnins.2012.00171}
}
\concept{ggseg3d_atlases}
