% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dkt}
\alias{dkt}
\title{Desikan-Killiany-Tourville Cortical Atlas}
\format{
ggseg brain-atlas
}
\usage{
data(dkt)
}
\description{
Coordinate data for the Desikan-Killany-Tourville Cortical atlas,
with 31 regions in on the cortical surface of the brain.
}
\examples{
data(dkt)
}
\references{
101 labeled brain images and a consistent human cortical labeling protocol. Arno Klein, Jason Tourville. Frontiers in Brain Imaging Methods. 6:171. DOI: 10.3389/fnins.2012.00171
(\href{https://doi.org/10.3389/fnins.2012.00171}{PubMed})
}
\concept{ggseg_atlases}
\keyword{datasets}
