% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dkt}
\alias{dkt}
\title{Desikan-Killiany-Tourville Cortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(dkt)
}
\description{
Brain atlas for the Desikan-Killiany-Tourville cortical parcellation
with 31 regions per hemisphere.
Contains both 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
data(dkt)
dkt
}
\references{
Klein A, Tourville J (2012). "101 labeled brain images and a
consistent human cortical labeling protocol." \emph{Frontiers in Neuroscience},
6:171. \doi{10.3389/fnins.2012.00171}
}
\concept{ggseg_atlases}
\keyword{datasets}
