% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dkt_3d}
\alias{dkt_3d}
\title{Desikan-Killiany-Tourville Cortical Atlas}
\format{
A tibble with 4 observations and a nested data.frame
\describe{
  \item{surf}{type of surface (`inflated` or `white`)}
  \item{hemi}{hemisphere (`left`` or `right`)}
  \item{data}{data.frame of necessary variables for plotting
  }

  \item{atlas}{String. atlas name}
  \item{roi}{numbered region from surface}
  \item{annot}{concatenated region name}
  \item{label}{label `hemi_annot` of the region}
  \item{mesh}{list of meshes in two lists: vb and it}
  \item{acronym}{abbreviated name of annot}
  \item{lobe}{lobe localization}
  \item{region}{name of region in full}
  \item{colour}{HEX colour of region}
}
}
\usage{
data(dkt_3d)
}
\description{
Coordinate data for the Desikan-Killany-Tourville Cortical atlas,
with 31 regions in on the cortical surface of the brain.
}
\examples{
data(dkt_3d)
}
\references{
101 labeled brain images and a consistent human cortical labeling protocol. Arno Klein, Jason Tourville. Frontiers in Brain Imaging Methods. 6:171. DOI: 10.3389/fnins.2012.00171
(\href{https://doi.org/10.3389/fnins.2012.00171}{PubMed})
}
\concept{ggseg3d_atlases}
\keyword{datasets}
